refactor(broker): remove AtomixLogStoragePartitionStep
This step built the AtomixLogStorage. However, this class is only used in one place, the LogStream. This change simplifies partition context and the transition steps.
