+++ broker/src/main/java/io/camunda/zeebe/broker/Broker.java
@@ -53,7 +53,6 @@ import io.camunda.zeebe.broker.system.partitions.ZeebePartition;
 import io.camunda.zeebe.broker.system.partitions.impl.AtomixPartitionMessagingService;
 import io.camunda.zeebe.broker.system.partitions.impl.PartitionProcessingState;
 import io.camunda.zeebe.broker.system.partitions.impl.PartitionTransitionImpl;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.AtomixLogStoragePartitionStep;
 import io.camunda.zeebe.broker.system.partitions.impl.steps.ExporterDirectorPartitionStep;
 import io.camunda.zeebe.broker.system.partitions.impl.steps.FollowerPostStoragePartitionStep;
 import io.camunda.zeebe.broker.system.partitions.impl.steps.LeaderPostStoragePartitionStep;
@@ -96,7 +95,6 @@ public final class Broker implements AutoCloseable {
   public static final Logger LOG = Loggers.SYSTEM_LOGGER;
   private static final List<PartitionStep> LEADER_STEPS =
       List.of(
-          new AtomixLogStoragePartitionStep(),
           new LogStreamPartitionStep(),
           new SnapshotReplicationPartitionStep(),
           new StateControllerPartitionStep(),
